Reorder schema properties according to significance (kind of)
Until now the properties appaered to be in a somewhat random order
which made it difficult to easily grasp what an object is about.
I tried to put the more significant properties (e.g. id, uri, name,
title) on the top of the list while sometimes resorting to
alphabetical order for the less significant ones.

fix