Remove fixed values for relation param and introduce relation-role param
We no longer want to restrict the kind of relations that can be modeled
for characters of a play to a fixed set. That makes it necessary, for
directed relations, to be able to select the active or passive side of
the relation. This is now possible with the `relation-role` parameter.

The old behavior of suffixing the relation name with either '_active' or
'_passive' is still supported.

See https://github.com/dracor-org/dracor-schema/discussions/80.